+++ kaithem/src/docs/changes.md
@@ -13,6 +13,8 @@ Change Log
 - Many bugfixes
 - Boot time is several times faster
 - No limit to event traceback stack depth
+- Remove posting to /system/threads/start, it created a refactoring nightmare and wasn't useful
+- Remove system/errors/workers for the same reason, traditional logging makes it obsolete.
 
 ### 0.63.1
 -  Fix JS dependancy error in lighting module